refactor: move resource/convert in ext
